Changed this to functions file as it was needed to allow pytest
